backend/requirements.txt
# =============================================================================
# CRI Chatbot Platform — Python Dependencies
# =============================================================================

# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1

# Redis
redis[hiredis]==5.2.1

# Vector Store
qdrant-client==1.12.1

# Object Storage
miniopy-async==1.21.1

# AI / LLM
google-genai>=1.0.0
tenacity>=9.0.0
tiktoken>=0.7.0
langchain==0.3.14
langgraph==0.2.60

# HTTP client
httpx==0.28.1

# Logging
structlog==24.4.0

# Security
cryptography>=43.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20

# Monitoring
prometheus-client==0.21.1
prometheus-fastapi-instrumentator==7.0.2

# Task queue
arq==0.26.1

# Document parsing (KB ingestion)
pdfplumber>=0.11.0
python-docx>=1.1.0
openpyxl>=3.1.0
reportlab>=4.1.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0

# Dev
ruff==0.8.6
mypy==1.14.1
pytest-cov
